finish all of operator of numberObject

+++ OOPFinalProject/NumberObject.cpp
@@ -10,7 +10,7 @@ NumberObject::NumberObject(const NumberObject& _numberObject) {
 	this->numData = _numberObject.numData;
 }
 
-NumberObject::NumberObject(int _number) {
+NumberObject::NumberObject(long long int _number) {
 	this->numType = INTEGER;
 	this->numData = Integer(_number).getNumData();
 }
